refactor(library/search): split 2596-line file into mod/auto_search/interactive/grab + tests
handlers/library/search.rs was the second-largest unsplit file in the
repo per the v1.5 large-files audit. Production code now lives in 4
cohesive siblings under handlers/library/search/:

  mod.rs         138  Slim orchestrator: anilist_search +
                      api_series_detail (the cheap API endpoints) plus
                      the public re-export surface main.rs reaches via
                      handlers::library::search::*. utoipa __path_*
                      symbols are re-exported alongside each function
                      so the OpenAPI schema generator still finds them.
  auto_search.rs 912  Auto-search pipeline + auto-expand sibling-pack
                      detection. Kept together because
                      auto_expand_library_from_pack and
                      run_auto_search_targets_with_upgrades call each
                      other in both directions; splitting them would
                      cross the module boundary on every call edge.
                      Owns auto_search_series, auto_search_episode,
                      run_auto_search_targets (now pub for episodes.rs
                      callers), AutoSearchQuery, and the per-handler
                      progress-event helpers.
  interactive.rs 398  search_batch_releases, interactive_search_episode,
                      interactive_search_batches.
  grab.rs        397  grab_batch_result, grab_interactive_result.

Tests stay in a single tests/mod.rs (806 LoC, just under the ~1500 LoC
topic-split threshold from CLAUDE.md). The auto_expand cumulative-
offset / sibling-routing test bodies didn't change — only their
imports got the explicit names that the prior `use super::*;`
implicitly inherited from the bigger search.rs.

Cross-sibling visibility: batch_episode_numbers,
auto_expand_library_from_pack, and display_title_for_progress are now
`pub(super)` so interactive + grab can call them; AutoSearchQuery and
run_auto_search_targets are `pub` (with `pub use` re-exports from
mod.rs) so external callers in handlers/library/{crud,episodes}.rs
keep working.

Verify: 1233 unit + 7 PR0 + 1 e2e tests all green, fmt + clippy
-D warnings clean.
